requirements/dev.txt
-r test.txt
-r docs.txt

django-extensions==2.1.6
django-debug-toolbar==1.11
